Completed exercise 4.21

+++ part4/controllers/blogs.js
@@ -63,7 +63,39 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    
+    if (!decodedToken.id) {
+        return response.status(401).json({
+            error: 'token missing or invalid'
+        })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(404).json({
+            error: 'The requested blog does not exist'
+        })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    if (!user) {
+        return response.status(404).json({
+            error: 'The user account no longer exists'
+        })
+    }
+
+    if (!(blog.user.toString() === user._id.toString())) {
+        return response.status(401).json({
+            error: 'Only the creator of the blog can delete it'
+        })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
+    user.blogs = user.blogs.filter(blog => blog.toString() !== request.params.id)
+    await user.save()
     response.status(204).end()
 })
